add day 49
Day 49 looks to be script to automatically apply for a job through linkedIn.com using selenium (in this case the inherited Selenium2 I created to handle some of the leg work). Inside the Selenium2, as before, we make use of the geckodriver to run the we browser of choice (firefox).