Add days 54-57
Day 54 is an intro to web development with Flask. Day 55 is a continutation of web development with backend development with a number guessing game. Day 56 starts with template web development in Flask, this is a resume web page. Day 57 further expands on template development using Jinja by building a blog page.