Add days 58-61
Day 58 is a crash course in HTML  and css writing a fictional "tindog" web page. Day 59 is a blog webpage with templates with Flask/Jinja. Day 60 is the login for the blog page using forms. Day 61 is a login form example with validation, displays separate pages for successful vs. unsuccessful login.