Day 85
A Tkinter GUI desktop application that tests your typing speed. I
have a master list of words I downloaded from a reading practice website of  various lenghts. A couple buttons to start/stop/reset the typing test. I use a separate thread to monitor typing times. When the length of word is matched or greater, the score is adjusted and a new word is retrieved.  The "word_list.tx_" is the wordlist, changed to "word_list.txt".