add day 90
python script to convert a .pdf file to an audio book. I found a webpage describing how to load .pdf files in python and merely needed to send text to Amazon's Polly to convert to speech. It's been a while, but I believe Amazon provided the polly.py code.

for extra credit, Amazon also provided a website to convert text in a textarea to speech in various voices. I used pdf.js to convert the loaded .pdf to text, insert into the text area and send to polly.

None of the voices returned have any inflections, or any variations in speech patterns as they're paid services at the time of the application implementation.